add keyboard shortcuts to tooltips
