+++ components/global/at_mentions_button/at_mentions_button.tsx
@@ -13,6 +13,8 @@ import OverlayTrigger from 'components/overlay_trigger';
 import {getIsRhsOpen, getRhsState} from 'selectors/rhs';
 import {GlobalState} from 'types/store';
 import Constants, {RHSStates} from 'utils/constants';
+import {KEYBOARD_SHORTCUTS} from '../../keyboard_shortcuts/keyboard_shortcuts';
+import KeyboardShortcutSequence from '../../keyboard_shortcuts/keyboard_shortcuts_sequence/keyboard_shortcuts_sequence';
 
 const AtMentionsButton = (): JSX.Element => {
     const dispatch = useDispatch();
@@ -33,6 +35,11 @@ const AtMentionsButton = (): JSX.Element => {
             <FormattedMessage
                 id='channel_header.recentMentions'
                 defaultMessage='Recent mentions'
+            />
+            <KeyboardShortcutSequence
+                shortcut={KEYBOARD_SHORTCUTS.navMentions}
+                hideDescription={true}
+                isInsideTooltip={true}
             />
         </Tooltip>
     );
