+++ components/global/history_buttons/history_buttons.tsx
@@ -3,11 +3,16 @@
 
 import React from 'react';
 import styled from 'styled-components';
+import {Tooltip} from 'react-bootstrap';
 import IconButton from '@mattermost/compass-components/components/icon-button';
 
 import {trackEvent} from 'actions/telemetry_actions';
 import * as Utils from 'utils/utils';
 import {browserHistory} from 'utils/browser_history';
+import {KEYBOARD_SHORTCUTS, KeyboardShortcutsSeq} from '../../keyboard_shortcuts/keyboard_shortcuts';
+import Constants from '../../../utils/constants';
+import KeyboardShortcutSequence from '../../keyboard_shortcuts/keyboard_shortcuts_sequence/keyboard_shortcuts_sequence';
+import OverlayTrigger from 'components/overlay_trigger';
 
 const HistoryButtonsContainer = styled.nav`
     display: flex;
@@ -19,6 +24,17 @@ const HistoryButtonsContainer = styled.nav`
 `;
 
 const HistoryButtons = (): JSX.Element => {
+    const getTooltip = (shortcut: KeyboardShortcutsSeq) => (
+        <Tooltip
+            id='upload-tooltip'
+        >
+            <KeyboardShortcutSequence
+                shortcut={shortcut}
+                hoistDescription={true}
+                isInsideTooltip={true}
+            />
+        </Tooltip>
+    );
     const goBack = () => {
         trackEvent('ui', 'ui_history_back');
         browserHistory.goBack();
@@ -31,22 +47,36 @@ const HistoryButtons = (): JSX.Element => {
 
     return (
         <HistoryButtonsContainer>
-            <IconButton
-                icon={'arrow-left'}
-                onClick={goBack}
-                size={'sm'}
-                compact={true}
-                inverted={true}
-                aria-label={Utils.localizeMessage('sidebar_left.channel_navigator.goBackLabel', 'Back')}
-            />
-            <IconButton
-                icon={'arrow-right'}
-                onClick={goForward}
-                size={'sm'}
-                compact={true}
-                inverted={true}
-                aria-label={Utils.localizeMessage('sidebar_left.channel_navigator.goForwardLabel', 'Forward')}
-            />
+            <OverlayTrigger
+                trigger={['hover']}
+                delayShow={Constants.OVERLAY_TIME_DELAY}
+                placement='bottom'
+                overlay={getTooltip(KEYBOARD_SHORTCUTS.browserChannelPrev)}
+            >
+                <IconButton
+                    icon={'arrow-left'}
+                    onClick={goBack}
+                    size={'sm'}
+                    compact={true}
+                    inverted={true}
+                    aria-label={Utils.localizeMessage('sidebar_left.channel_navigator.goBackLabel', 'Back')}
+                />
+            </OverlayTrigger>
+            <OverlayTrigger
+                trigger={['hover']}
+                delayShow={Constants.OVERLAY_TIME_DELAY}
+                placement='bottom'
+                overlay={getTooltip(KEYBOARD_SHORTCUTS.browserChannelNext)}
+            >
+                <IconButton
+                    icon={'arrow-right'}
+                    onClick={goForward}
+                    size={'sm'}
+                    compact={true}
+                    inverted={true}
+                    aria-label={Utils.localizeMessage('sidebar_left.channel_navigator.goForwardLabel', 'Forward')}
+                />
+            </OverlayTrigger>
         </HistoryButtonsContainer>
     );
 };
