+++ components/keyboard_shortcuts/keyboard_shortcuts_sequence/keyboard_shortcuts_sequence.scss
@@ -1,0 +1,30 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+.shortcut-line {
+    margin: 16px 0;
+
+    span {
+        &:first-child {
+            margin-right: 5px;
+        }
+    }
+
+    .shortcut-key {
+        padding: 2px 5px;
+
+        + .shortcut-key {
+            margin-left: 2px;
+        }
+    }
+}
+
+.tooltip-inner .shortcut-line {
+    margin: 4px 0;
+}
+
+@media (max-height: 800px) {
+    .shortcut-line {
+        margin-top: 14px;
+    }
+}
