+++ components/shortcut_key/shortcut_key.tsx
@@ -6,6 +6,8 @@ import './shortcut_key.scss';
 
 export enum ShortcutKetVariant {
     Contrast = 'contrast',
+    Tooltip = 'tooltip',
+    ShortcutModal = 'shortcut',
 }
 
 export type ShortcutKeyProps = {
@@ -17,6 +19,10 @@ export const ShortcutKey = ({children, variant}: ShortcutKeyProps) => {
     let className = 'shortcut-key';
     if (variant === ShortcutKetVariant.Contrast) {
         className += ' shortcut-key--contrast';
+    } else if (variant === ShortcutKetVariant.Tooltip) {
+        className += ' shortcut-key--tooltip';
+    } else if (variant === ShortcutKetVariant.ShortcutModal) {
+        className += ' shortcut-key--shortcut-modal';
     }
 
     return (
