+++ i18n/en.json
@@ -4298,8 +4298,8 @@
   "team.button.ariaLabel": "{teamName} team",
   "team.button.mentions.ariaLabel": "{teamName} team, {mentionCount} mentions",
   "team.button.name_undefined": "This team does not have a name",
-  "team.button.tooltip": "Ctrl+Alt+{order}",
-  "team.button.tooltip.mac": "⌘ ⌥ {order}",
+  "team.button.tooltip": "Ctrl|Alt|{order}",
+  "team.button.tooltip.mac": "⌘|⌥|{order}",
   "team.button.unread.ariaLabel": "{teamName} team unread",
   "terms_of_service.agreeButton": "I Agree",
   "terms_of_service.api_error": "Unable to complete the request. If this issue persists, contact your System Administrator.",
