+++ sass/routes/_module.scss
@@ -11,6 +11,5 @@
 @import 'loading';
 @import 'print';
 @import 'settings';
-@import 'shortcuts-modal';
 @import 'signup';
 @import 'statistics';
